Customer Statements report added

+++ retail-ui/src/main/webapp/scripts/cimg/logoutController.js
@@ -4,10 +4,17 @@
 
 cimgApp.controller('logoutController', function($http, $scope, UserService,multiPageService, $state, LOGOUT_URI, baseDataService) {
     //$scope.userService = UserService;
-    baseDataService.getBaseData(LOGOUT_URI).then(function(response){
+    baseDataService.displayMessage('yesNo','Warning!!','Are you sure you want to logout?').then(function(result){
+        if (result) {
+            baseDataService.getBaseData(LOGOUT_URI).then(function(response){
+            });
+            UserService.removeUser();
+            multiPageService.removePageList();
+            $state.go('dashboard.login');
+        } else {
+            return;
+        }
     });
-    UserService.removeUser();
-    multiPageService.removePageList();
-    $state.go('dashboard.login');
+
 })
